LPS-104983 Dispatch network states when using serviceFetch
Dispatching network status by services is still a bit manual, so every new thunk or service will have to pass the `onNetworkStatus` callback.

If this becomes a problem in the future we can go with implementing an Observer that will create a serviceFetch communication channel with the Store, minimizing the use of callbacks. I have not done this abstraction now to allow us to test and hope that everyone agrees.