LPS-105011 configuration-admin-web: Isolate UID handling to better show this is a Non Standard UID (ConfigurationModel is not a ClassedModel).
With the introduction of Low Level Search, this class can be simplified
into a plain DAO pattern with a separate index for system settings
metadata (instead of retrofitting the Indexer framework's complexity to
operate a liferay-0 index that holds no user content). This being done,
document ID can then be delegated to the search engine.

https://issues.liferay.com/browse/LPS-105011